Move 'addTemplateField' component in separated JS file
