+++ src/ggrc-client/js/components/assessment_attributes.js
@@ -201,114 +201,4 @@
       },
     },
   });
-
-  GGRC.Components('addTemplateField', {
-    tag: 'add-template-field',
-    template: can.view(GGRC.mustache_path +
-      '/assessment_templates/attribute_add_field.mustache'),
-    scope: function (attrs, parentScope) {
-      return new can.Map({
-        selected: new can.Map(),
-        fields: parentScope.attr('fields'),
-        types: parentScope.attr('types'),
-        // the field types that require a list of possible values to be defined
-        valueAttrs: ['Dropdown'],
-        /*
-         * Create a new field.
-         *
-         * Field must contain value title, type, values.
-         * Opts are populated, once we start changing checkbox values
-         *
-         * @param {can.Map} scope - the current (add-template-field) scope
-         * @param {jQuery.Object} el - the clicked DOM element
-         * @param {Object} ev - the event object
-         */
-        addField: function (scope, el, ev) {
-          let fields = scope.attr('fields');
-          let selected = scope.attr('selected');
-          let title = _.trim(selected.title);
-          let type = _.trim(selected.type);
-          let invalidInput = false;
-          let values = _.splitTrim(selected.values, {
-            unique: true,
-          }).join(',');
-          ev.preventDefault();
-          scope.attr('selected.invalidTitle', false);
-          scope.attr('selected.emptyTitle', false);
-          scope.attr('selected.dublicateTitle', false);
-          scope.attr('selected.invalidValues', false);
-
-          if (this.isEmptyTitle(title)) {
-            this.attr('selected.invalidTitle', true);
-            this.attr('selected.emptyTitle', true);
-            invalidInput = true;
-          } else if (this.isDublicateTitle(fields, title)) {
-            this.attr('selected.invalidTitle', true);
-            this.attr('selected.dublicateTitle', true);
-            invalidInput = true;
-          }
-          if (this.isInvalidValues(scope.valueAttrs, type, values)) {
-            scope.attr('selected.invalidValues', true);
-            invalidInput = true;
-          }
-          if (invalidInput) {
-            return;
-          }
-          // We need to defer adding in modal since validation is preventing
-          // on adding the first item
-          _.defer(function () {
-            fields.push({
-              id: scope.attr('id'),
-              title: title,
-              attribute_type: type,
-              multi_choice_options: values,
-            });
-            _.each(['title', 'values', 'multi_choice_options'],
-              function (type) {
-                selected.attr(type, '');
-              });
-          });
-        },
-        isInvalidValues: function (valueAttrs, type, values) {
-          return _.contains(valueAttrs, type) && !values;
-        },
-        isDublicateTitle: function (fields, selectedTitle) {
-          let duplicateField = _.some(fields, function (item) {
-            return item.title === selectedTitle && !item._pending_delete;
-          });
-          return fields.length && duplicateField;
-        },
-        isEmptyTitle: function (selectedTitle) {
-          return !selectedTitle;
-        },
-      });
-    },
-    events: {
-      /*
-       * Set default dropdown type on init
-       */
-      init: function () {
-        let types = this.scope.attr('types');
-        if (!this.scope.attr('selected.type')) {
-          this.scope.attr('selected.type', _.first(types).attr('type'));
-        }
-      },
-    },
-    helpers: {
-      /*
-       * Get input placeholder value depended on type
-       *
-       * @param {Object} options - Mustache options
-       */
-      placeholder: function (options) {
-        let types = this.attr('types');
-        let item = _.findWhere(types, {
-          type: this.attr('selected.type'),
-        });
-        if (item) {
-          return item.text;
-        }
-      },
-    },
-  });
 })(window.can, window.can.$);
