+++ src/ggrc-client/js/components/assessment_templates/templates/add-template-field.mustache
